+++ tests/ExchangeRateProvider/BaseCurrencyProviderTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Brick\Money\Tests\ExchangeRateProvider;
 
+use Brick\Math\BigNumber;
 use Brick\Money\Currency;
 use Brick\Money\Exception\ExchangeRateProviderException;
 use Brick\Money\ExchangeRateProvider;
@@ -75,9 +76,15 @@ class BaseCurrencyProviderTest extends AbstractTestCase
 
     public function testThrowsExceptionWhenReverseRateIsZero(): void
     {
-        $provider = ConfigurableProvider::builder()
-            ->addExchangeRate('USD', 'EUR', '0')
-            ->build();
+        $provider = new class() implements ExchangeRateProvider {
+            public function getExchangeRate(Currency $sourceCurrency, Currency $targetCurrency, array $dimensions = []): ?BigNumber
+            {
+                return match ([$sourceCurrency->getCurrencyCode(), $targetCurrency->getCurrencyCode()]) {
+                    ['USD', 'EUR'] => BigNumber::of('0'),
+                    default => null,
+                };
+            }
+        };
 
         $baseProvider = new BaseCurrencyProvider($provider, 'USD');
 
@@ -89,10 +96,16 @@ class BaseCurrencyProviderTest extends AbstractTestCase
 
     public function testThrowsExceptionWhenSourceRateIsZero(): void
     {
-        $provider = ConfigurableProvider::builder()
-            ->addExchangeRate('USD', 'EUR', '0')
-            ->addExchangeRate('USD', 'GBP', '0.8')
-            ->build();
+        $provider = new class() implements ExchangeRateProvider {
+            public function getExchangeRate(Currency $sourceCurrency, Currency $targetCurrency, array $dimensions = []): ?BigNumber
+            {
+                return match ([$sourceCurrency->getCurrencyCode(), $targetCurrency->getCurrencyCode()]) {
+                    ['USD', 'EUR'] => BigNumber::of('0'),
+                    ['USD', 'GBP'] => BigNumber::of('0.8'),
+                    default => null,
+                };
+            }
+        };
 
         $baseProvider = new BaseCurrencyProvider($provider, 'USD');