+++ tests/RationalMoneyTest.php
@@ -15,6 +15,7 @@ use Brick\Money\Context\CustomContext;
 use Brick\Money\Context\DefaultContext;
 use Brick\Money\Currency;
 use Brick\Money\Exception\CurrencyMismatchException;
+use Brick\Money\Exception\InvalidArgumentException;
 use Brick\Money\Exception\UnknownCurrencyException;
 use Brick\Money\Money;
 use Brick\Money\RationalMoney;
@@ -352,6 +353,8 @@ class RationalMoneyTest extends AbstractTestCase
             [['3/7', 'USD'], 'EUR', '7/3', 'EUR 1'],
             [['100', 'USD'], 'JPY', 150, 'JPY 15000'],
             [['1.23', 'USD'], 'INVALID', '1', UnknownCurrencyException::class],
+            [['1.23', 'USD'], 'EUR', '0', InvalidArgumentException::class],
+            [['1.23', 'USD'], 'EUR', '-1', InvalidArgumentException::class],
         ];
     }